requirements.txt
fastapi
uvicorn
cassandra-driver
pydantic-settings
python-dotenv